Final Commit For Brute Force Building
The world was building up to this point, but as this project is being built by following a lecture so far, it's time to replace the project with a new one, as the course instructor suggests. This version probably won't work without a significant amount of restoration to fix compile erros.